Add robots meta tag for staging environment
Both, staging and production environment are currently indexed. This
creates a duplicate content¹ problem. This can be solved in two
different ways. Using the robots meta tag² or using canonical³ URLs.
In your specific case there is no legitimate reason to have the staging
environment indexed so a "noindex" is appropriate.

[1] https://support.google.com/webmasters/answer/66359
[2] https://developers.google.com/webmasters/control-crawl-index/docs/robots_meta_tag
[3] https://support.google.com/webmasters/answer/139066