Read binary even text in Python
according http://docs.python-requests.org/en/master/user/quickstart/#post-a-multipart-encoded-file 

'It is strongly recommended that you open files in binary mode. This is because Requests may attempt to provide the Content-'Length header for you, and if it does this value will be set to the number of bytes in the file. Errors may occur if you open the file 'in text mode.