Remove mailgun-go NewMessage function public key parameter
[Issue 72](https://github.com/mailgun/mailgun-go/issues/72)

It looks like you have changed the public going from V1 to V2 in that the NewMailgun function no longer takes a publicApi parameter. I have updated the Go code examples that have not been updated to reflect this. I don't see a separation of documentation for V1 and V2.