Update User Manual > Securing Webhooks
The instructions for checking the authenticity of a webhook message previously instructed using the "API Key" as the SHA HMAC key, however when interpreting this as the "Private API key" under Settings > API Keys, this results in validation failing. Only after some time struggling and experimenting did I discover that the expected key is the "HTTP webhook signing key", not the "Private API key".

This commit updates the "Securing Webhooks" instructions to use the "Webhook Signing Key" (as it is labeled across the Mailgun web app) to make it clearer which key should be used.